+++ src/2015/buri/aoc/y15/d25/Day25.java
@@ -44,9 +44,9 @@ public class Day25 extends BasePuzzle {
 	 * Calculates the next position to insert a code. X is row count (vertical) and Y is column count (horizontal).
 	 */
 	private static void move(Pair pair) {
-		pair.setX(pair.getX() - 1);
-		pair.setY(pair.getY() + 1);
-		if (pair.getX() == 0) {
+		pair.setX(pair.getX().intValue() - 1);
+		pair.setY(pair.getY().intValue() + 1);
+		if (pair.getX().intValue() == 0) {
 			pair.setX(pair.getY());
 			pair.setY(1);
 		}
