+++ src/2016/buri/aoc/y16/d13/Grid.java
@@ -106,17 +106,17 @@ public class Grid extends CharGrid {
 	 */
 	private List<Pair> getTraversableNeighbors(Pair center) {
 		List<Pair> neighbors = new ArrayList<>();
-		if (center.getY() > 0) {
-			neighbors.add(new Pair(center.getX(), center.getY() - 1));
+		if (center.getY().intValue() > 0) {
+			neighbors.add(new Pair(center.getX(), center.getY().intValue() - 1));
 		}
-		if (center.getX() > 0) {
-			neighbors.add(new Pair(center.getX() - 1, center.getY()));
+		if (center.getX().intValue() > 0) {
+			neighbors.add(new Pair(center.getX().intValue() - 1, center.getY()));
 		}
-		if (center.getX() < getWidth() - 1) {
-			neighbors.add(new Pair(center.getX() + 1, center.getY()));
+		if (center.getX().intValue() < getWidth() - 1) {
+			neighbors.add(new Pair(center.getX().intValue() + 1, center.getY()));
 		}
-		if (center.getY() < getHeight() - 1) {
-			neighbors.add(new Pair(center.getX(), center.getY() + 1));
+		if (center.getY().intValue() < getHeight() - 1) {
+			neighbors.add(new Pair(center.getX(), center.getY().intValue() + 1));
 		}
 		// Remove any that are not traversable.
 		for (Iterator<Pair> iterator = neighbors.iterator(); iterator.hasNext();) {
