+++ src/2017/buri/aoc/y17/d19/Diagram.java
@@ -79,10 +79,10 @@ public class Diagram extends CharGrid {
 	 * Determines the next direction at a corner. Assumes 1 true path and always 90 degrees.
 	 */
 	private Direction getNextDirection() {
-		Pair up = new Pair(getCurrentPosition().getX(), getCurrentPosition().getY() - 1);
-		Pair right = new Pair(getCurrentPosition().getX() + 1, getCurrentPosition().getY());
-		Pair down = new Pair(getCurrentPosition().getX(), getCurrentPosition().getY() + 1);
-		Pair left = new Pair(getCurrentPosition().getX() - 1, getCurrentPosition().getY());
+		Pair up = new Pair(getCurrentPosition().getX(), getCurrentPosition().getY().intValue() - 1);
+		Pair right = new Pair(getCurrentPosition().getX().intValue() + 1, getCurrentPosition().getY());
+		Pair down = new Pair(getCurrentPosition().getX(), getCurrentPosition().getY().intValue() + 1);
+		Pair left = new Pair(getCurrentPosition().getX().intValue() - 1, getCurrentPosition().getY());
 		Direction next = null;
 		if (getCurrentDirection() != null) {
 			switch (getCurrentDirection()) {
@@ -113,8 +113,8 @@ public class Diagram extends CharGrid {
 	 * Returns true if position is within the diagram.
 	 */
 	private boolean isBounded(Pair position) {
-		boolean xBounded = (position.getX() >= 0 && position.getX() < getWidth());
-		boolean yBounded = (position.getY() >= 0 && position.getY() < getHeight());
+		boolean xBounded = (position.getX().intValue() >= 0 && position.getX().intValue() < getWidth());
+		boolean yBounded = (position.getY().intValue() >= 0 && position.getY().intValue() < getHeight());
 		return (xBounded && yBounded);
 	}
