+++ src/2019/buri/aoc/y19/d15/Maze.java
@@ -160,10 +160,10 @@ public class Maze extends CharGrid {
 	 */
 	private List<Pair> getTraversableNeighbors(Pair center, Pair origin) {
 		List<Pair> neighbors = new ArrayList<>();
-		neighbors.add(new Pair(center.getX(), center.getY() - 1));
-		neighbors.add(new Pair(center.getX() - 1, center.getY()));
-		neighbors.add(new Pair(center.getX() + 1, center.getY()));
-		neighbors.add(new Pair(center.getX(), center.getY() + 1));
+		neighbors.add(new Pair(center.getX(), center.getY().intValue() - 1));
+		neighbors.add(new Pair(center.getX().intValue() - 1, center.getY()));
+		neighbors.add(new Pair(center.getX().intValue() + 1, center.getY()));
+		neighbors.add(new Pair(center.getX(), center.getY().intValue() + 1));
 		// Remove any that are not traversable.
 		for (Iterator<Pair> iterator = neighbors.iterator(); iterator.hasNext();) {
 			Pair position = iterator.next();
